styles move to styled.ts
